requirements.txt
tensorflow
scikit-learn
tqdm
pillow
numpy
